Created the 3 level scripts
The LevelEntrance Script contains the ID (string) and a bool to see if the entrance is a default entrance. The LevelExit contains target ID (string) of place to spawn in and another string for the scene to spawn in. It also contains an empty OnTriggerEnter method. The LevelLoader is unfinished