Jonny's branch (#119)
* FootStep Script

Script to play an event for every step. I.E. footstep sound

* Footstep source script

* Created the 3 level scripts

The LevelEntrance Script contains the ID (string) and a bool to see if the entrance is a default entrance. The LevelExit contains target ID (string) of place to spawn in and another string for the scene to spawn in. It also contains an empty OnTriggerEnter method. The LevelLoader is unfinished

* Comments for the LevelLoader script

* LevelEntrance and LevelExit got a bit tweeked and LevelLoader was completed

* Added comments to the scripts

* created test scene and added a timing thing for entrances

* style refactoring

---------

Co-authored-by: Daniel Walls <walls.dk@gmail.com>