WedgeOfTreesArena and Minor Additions to DeforestedArea (#309)
* Added tree decor to DeforestedArea

Changed the type of trees in DeforestedArea to a few different varieties, added logs and stumps to the surrounding area. Also few minor tweaks to walls, terrain, the camera trigger, and foliage.

* Yarn and Combat Triggers to Deforested

Added a placeholder yarn script for interacting with Ghost-a-light's cage, and created triggers for encounters which still need to be set up.

* Started new forest arena: Wedge of Trees

Just the creation of a new arena scene with the sub arena spawn logic. Not currently usable without putting in build settings, and still debating how to set up terrain.

* Added a new arena: A Wedge of Trees

Have the basic shape for the arena with spawners for enemies and players, but would need to put in build settings to see if it actually works.

* Added Wedge of Trees arena to build settings

Made some adjustments to the Wedge of Trees arenas to shrink the playing area  and added it to build settings so a transition to it from deforested area could work.

---------

Co-authored-by: Daniel Walls <walls.dk@gmail.com>