Jonny's new branch (#331)
* SuperJump

Script

* Updated Jump Script with a few things

Added indicator for mouse

* WindPush script

Most of it was completed according to the GitHub description, Damage dealt has not been applied though. Still needs to be tested

* meta files + player action

* style changes

* Started the Ghost-A-Light Brain script

Started script. Added feature of turning invisible when a player monster is too close. Need to add logic for when an enemy is not too close. Also, need to fix a few bugs

* GhostALight Brain

Made a few comments on how to do fix it.

* GhostALight Brain script finished

Brain script is finished and was tested for the most part.

---------

Co-authored-by: Daniel Walls <walls.dk@gmail.com>